+++ rotkehlchen/tests/utils/ethereum.py
@@ -1,7 +1,8 @@
+from rotkehlchen.chain.ethereum.manager import NodeName
+
 ETHEREUM_TEST_PARAMETERS = ['ethrpc_endpoint,ethereum_manager_connect_at_start', [
-    # Query etherscan
-    ('', False),
-    # For real node querying use infura for now since blockscout is having server problems
-    # ('https://mainnet-nethermind.blockscout.com', True),
-    ('https://mainnet.infura.io/v3/66302b8fb9874614905a3cbe903a0dbb', True),
+    # Query etherscan only
+    ('', ()),
+    # For "our own" node querying use infura
+    ('https://mainnet.infura.io/v3/66302b8fb9874614905a3cbe903a0dbb', (NodeName.OWN,)),
 ]]